fix: rename logging module
